rename event peek -> event context
