+++ lib/logflare_web/live/search_live/templates/logs_search.html.heex
@@ -14,7 +14,6 @@
   ) %>
 <% end %>
 
-<%= live_modal(LogflareWeb.SearchLive.EventPeekComponent, id: "Event Peek", user: @user, source: @source, title: "Peek", return_to: nil, event_id: "e9782bbc-5f19-4c60-9ea2-15f1b304960d") %>
 <div id="source-logs-search-control" class="subhead " phx-hook="SourceLogsSearch">
   <div class="container mx-auto">
     <h5>
